# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Seletivo Seplag Gov-MT",
        "version": "0.1"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Autenticar usuário",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Informe o email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "teste@seplag.mt.gov.br"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Informe a senha",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "seplag2025"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successo no Login"
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/api/refresh": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Renovar Token",
                "operationId": "2c27d311a2876560fcb5a754b556cdc5",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Informe um email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "teste@seplag.mt.gov.br"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Insira a senha",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "seplag2025"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successo no Login"
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cidade": {
            "get": {
                "tags": [
                    "Cidades"
                ],
                "summary": "Mostra as Cidades cadastradas",
                "description": "Lista as Cidades",
                "operationId": "1eec1560cb76079fb0c015e722618860",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Nº da página",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cidades"
                ],
                "summary": "Cadastra uma nova Cidade",
                "description": "Registro de uma nova Cidade",
                "operationId": "a54ec416b780e0b13c71a29f0622f3c8",
                "parameters": [
                    {
                        "name": "cid_nome",
                        "in": "query",
                        "description": "Nome da cidade",
                        "required": true
                    },
                    {
                        "name": "cid_uf",
                        "in": "query",
                        "description": "UF",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Cidade não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cidade/{cid_id}": {
            "get": {
                "tags": [
                    "Cidades"
                ],
                "summary": "Mostra uma Cidade",
                "description": "Pesquisa por uma cidade através do (cid_id)",
                "operationId": "d50c61e536407329b142a652dd041baf",
                "parameters": [
                    {
                        "name": "cid_id",
                        "in": "path",
                        "description": "Nº de identificação da cidade",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cidade Encontrada",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Cidade não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cidades"
                ],
                "summary": "Exclui uma Cidade",
                "description": "Exclui uma Cidde através do (cid_id)",
                "operationId": "e9ed713c0e97583a24b97b0f3e835002",
                "parameters": [
                    {
                        "name": "cid_id",
                        "in": "path",
                        "description": "ID da Cidade",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cidade excluída com sucesso",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Não foi possível excluir a Cidade"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cidade/{ci_id}": {
            "put": {
                "tags": [
                    "Cidades"
                ],
                "summary": "Atualiza Cidade",
                "description": "Atualiza os dados de uma Cidade",
                "operationId": "4e4499e98d101b89a9ce935f1a7ada3f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cid_id",
                                    "cid_nome",
                                    "cid_uf"
                                ],
                                "properties": {
                                    "cid_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "cid_nome": {
                                        "type": "string",
                                        "example": "Pirinópolis"
                                    },
                                    "cid_uf": {
                                        "type": "string",
                                        "example": "MT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cidade atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cidade atualizada com sucesso"
                                        },
                                        "cidade": {
                                            "properties": {
                                                "cid_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "cid_nome": {
                                                    "type": "string",
                                                    "example": "Pirinópolis"
                                                },
                                                "cid_uf": {
                                                    "type": "string",
                                                    "example": "MT"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição inválida"
                    },
                    "404": {
                        "description": "Cidade não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}